PMM-11537-expose-toggle: move expose under switch component
