+++ public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/DBClusterAdvancedOptions/DBClusterAdvancedOptions.messages.ts
@@ -4,7 +4,6 @@ export const Messages = {
     pxcConfiguration: 'MySQL Configurations',
     mongodbConfiguration: 'MongoDB Configurations',
     commonConfiguration: 'Database Configurations',
-    networkAndSecurity: 'Network and Security',
   },
   labels: {
     nodes: 'Number of Nodes',
@@ -16,9 +15,6 @@ export const Messages = {
     pxcConfiguration: 'MySQL Configuration',
     mongodbConfiguration: 'MongoDB Configuration',
     commonConfiguration: 'Database Configuration',
-    expose: 'Expose',
-    internetFacing: 'Internet Facing',
-    sourceRange: 'Source Range',
   },
   resources: {
     small: 'Small',
@@ -26,17 +22,7 @@ export const Messages = {
     large: 'Large',
     custom: 'Custom',
   },
-
-  buttons: {
-    addNew: 'Add new',
-  },
-
-  tooltips: {
-    expose:
-      'You will make this database cluster available to connect from the internet. To limit access you need to specify source ranges',
-  },
   placeholders: {
     storageClass: 'storage class',
-    sourceRange: '181.170.213.40/32',
   },
 };
