--- public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/DBClusterAdvancedOptions/NetworkAndSecurity/NetworkAndSecurity.types.ts
@@ -1,5 +1,0 @@
-export enum NetworkAndSecurityFields {
-  expose = 'expose',
-  internetFacing = 'internetFacing',
-  sourceRanges = 'sourceRanges',
-}
