+++ public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/DBClusterBasicOptions/DBClusterBasicOptions.tsx
@@ -9,8 +9,6 @@ import {
   SelectFieldAdapter,
 } from 'app/percona/shared/components/Form/FieldAdapters/FieldAdapters';
 
-import { useSelector } from '../../../../../../types';
-import { getPerconaSettings } from '../../../../../shared/core/selectors';
 import { Kubernetes, Operator } from '../../../Kubernetes/Kubernetes.types';
 import { getDatabaseOptionFromOperator } from '../../../Kubernetes/Kubernetes.utils';
 import { KubernetesOperatorStatus } from '../../../Kubernetes/OperatorStatusItem/KubernetesOperatorStatus/KubernetesOperatorStatus.types';
@@ -28,7 +26,6 @@ import {
   Operators,
 } from './DBClusterBasicOptions.types';
 import { getKubernetesOptions, kubernetesClusterNameValidator, optionRequired } from './DBClusterBasicOptions.utils';
-import Restore from './Restore/Restore';
 
 const getAvailableDatabaseOptions = (kubernetesCluster: Kubernetes): DatabaseOption[] => {
   const { operators } = kubernetesCluster;
@@ -46,7 +43,6 @@ const getAvailableDatabaseOptions = (kubernetesCluster: Kubernetes): DatabaseOpt
 
 export const DBClusterBasicOptions: FC<DBClusterBasicOptionsProps> = ({ kubernetes, form }) => {
   const styles = useStyles(getStyles);
-  const { result: settings } = useSelector(getPerconaSettings);
   const { required, maxLength } = validators;
   const { change } = form;
   const { kubernetesCluster, databaseType } = form.getState().values;
@@ -127,7 +123,6 @@ export const DBClusterBasicOptions: FC<DBClusterBasicOptionsProps> = ({ kubernet
         label={Messages.labels.clusterName}
         validators={[required, kubernetesClusterNameValidator, maxLength(CLUSTER_NAME_MAX_LENGTH)]}
       />
-      {settings?.backupEnabled && <Restore form={form} />}
     </div>
   );
 };
