+++ public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/EditDBClusterPage.styles.ts
@@ -12,4 +12,14 @@ export const getStyles = ({ spacing }: GrafanaTheme2) => ({
     max-width: 170px;
     margin: ${spacing(6)} 0 ${spacing(3)} 0;
   `,
+
+  switchOptionsWrapper: css`
+    fieldset {
+      margin-top: 0;
+      margin-bottom: 0;
+      legend:first-child {
+        margin-bottom: ${spacing(2)};
+      }
+    }
+  `,
 });
