+++ public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/EditDBClusterPage.types.ts
@@ -7,25 +7,22 @@ import { Kubernetes } from '../../Kubernetes/Kubernetes.types';
 
 import { ConfigurationFields } from './DBClusterAdvancedOptions/Configurations/Configurations.types';
 import { AdvancedOptionsFields, DBClusterResources } from './DBClusterAdvancedOptions/DBClusterAdvancedOptions.types';
-import { NetworkAndSecurityFields } from './DBClusterAdvancedOptions/NetworkAndSecurity/NetworkAndSecurity.types';
 import { BasicOptionsFields, BasicOptionsFieldsProps } from './DBClusterBasicOptions/DBClusterBasicOptions.types';
 import { DBaaSBackupProps } from './DBaaSBackups/DBaaSBackups.types';
+import { NetworkAndSecurityProps } from './NetworkAndSecurity/NetworkAndSecurity.types';
 export type DBClusterPageMode = 'create' | 'edit' | 'list';
 
 export interface EditDBClusterPageProps {
   kubernetes: Kubernetes[];
 }
 
-export interface DBClusterCommonFormValues {
+export interface DBClusterCommonFormValues extends NetworkAndSecurityProps {
   [AdvancedOptionsFields.nodes]: number;
   [AdvancedOptionsFields.memory]: number;
   [AdvancedOptionsFields.cpu]: number;
   [AdvancedOptionsFields.disk]: number;
   [ConfigurationFields.configuration]?: string;
   [ConfigurationFields.storageClass]?: SelectableValue<string>;
-  [NetworkAndSecurityFields.expose]?: boolean;
-  [NetworkAndSecurityFields.sourceRanges]?: Array<{}> | [];
-  [NetworkAndSecurityFields.internetFacing]?: boolean;
 }
 export interface AddDBClusterFormValues
   extends ScheduledSectionFieldsValuesProps,
