+++ public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/NetworkAndSecurity/NetworkAndSecurity.messages.tsx
@@ -1,0 +1,19 @@
+export const Messages = {
+  labels: {
+    internetFacing: 'Internet Facing',
+    sourceRange: 'Source Range',
+  },
+  fieldSets: {
+    expose: 'Expose',
+  },
+  buttons: {
+    addNew: 'Add new',
+  },
+  tooltips: {
+    expose:
+      'You will make this database cluster available to connect from the internet. To limit access you need to specify source ranges',
+  },
+  placeholders: {
+    sourceRange: '181.170.213.40/32',
+  },
+};
