+++ public/app/percona/dbaas/components/DBCluster/EditDBClusterPage/NetworkAndSecurity/NetworkAndSecurity.types.ts
@@ -1,0 +1,11 @@
+export enum NetworkAndSecurityFields {
+  expose = 'expose',
+  internetFacing = 'internetFacing',
+  sourceRanges = 'sourceRanges',
+}
+
+export interface NetworkAndSecurityProps {
+  [NetworkAndSecurityFields.expose]?: boolean;
+  [NetworkAndSecurityFields.sourceRanges]?: Array<{ sourceRange: string }> | [];
+  [NetworkAndSecurityFields.internetFacing]?: boolean;
+}
